started Test view
