+++ server/models/tests.model.js
@@ -6,7 +6,7 @@ const TestSchema = new mongoose.Schema({
   marks: { type: Number, required: true },
   questions: { type: Array, required: true },
   rules: { type: Array, required: true },
-  scores: { type: Array, required: true },
+  scores: { type: Array, required: true, default:[] },
   duration: { type: String, required: true },
 });
